catalyst-backend/appsail/requirements.txt
# ── Core Framework ──
flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0

# ── Zoho Catalyst SDK ──
zcatalyst-sdk==0.0.2

# ── Image Processing ──
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.4

# ── AI Integration (Cloudflare Workers AI / Meta Llama 3) ──
requests==2.31.0

# ── Utilities ──
orjson==3.9.10
python-dotenv==1.0.0
